refactor: Move CalendarType enum to DateContext
