+++ components/ScheduleCalendar.tsx
@@ -4,9 +4,8 @@ import React, { useEffect } from 'react';
 import styles from './ScheduleCalendar.module.scss';
 
 import DayinSchedule from '@components/DayinSchedule';
-import { CalendarType } from '@components/Header/CalendarTypeDropDown';
 import Sidebar from '@components/Sidebar';
-import { useDateContext } from '@contexts/DateContext';
+import { CalendarType, useDateContext } from '@contexts/DateContext';
 import { useSidebarContext } from '@contexts/SidebarContext';
 
 export default function ScheduleCalendar() {
