+++ contexts/DateContext.tsx
@@ -8,15 +8,25 @@ import React, {
     useState,
 } from 'react';
 
+export enum CalendarType {
+    index = 'index',
+    day = 'day',
+    week = 'week',
+    month = 'month',
+    schedule = 'schedule',
+}
+
 interface DateContextData {
     yearNow: number;
     monthNow: number;
     dateNow: number;
     dayNow: number;
+    calendarType: CalendarType;
     setYearNow: Dispatch<SetStateAction<number>>;
     setMonthNow: Dispatch<SetStateAction<number>>;
     setDateNow: Dispatch<SetStateAction<number>>;
     setDayNow: Dispatch<SetStateAction<number>>;
+    setCalendarType: Dispatch<SetStateAction<CalendarType>>;
 }
 
 const DateContext = createContext<DateContextData>({
@@ -24,6 +34,7 @@ const DateContext = createContext<DateContextData>({
     monthNow: 0,
     dateNow: 0,
     dayNow: 0,
+    calendarType: CalendarType.index,
     setYearNow() {
         throw new Error('DateContext not provided');
     },
@@ -36,6 +47,9 @@ const DateContext = createContext<DateContextData>({
     setDayNow() {
         throw new Error('DateContext not provided');
     },
+    setCalendarType() {
+        throw new Error('DateContext not provided');
+    },
 });
 
 export const useDateContext = () => useContext(DateContext);
@@ -46,6 +60,7 @@ function DateProvider({ children }: PropsWithChildren) {
     const [monthNow, setMonthNow] = useState(now.getMonth() + 1);
     const [dateNow, setDateNow] = useState(now.getDate());
     const [dayNow, setDayNow] = useState(now.getDay());
+    const [calendarType, setCalendarType] = useState(CalendarType.month);
 
     const value = useMemo(
         () => ({
@@ -57,8 +72,10 @@ function DateProvider({ children }: PropsWithChildren) {
             setDateNow,
             dayNow,
             setDayNow,
+            calendarType,
+            setCalendarType,
         }),
-        [yearNow, monthNow, dateNow, dayNow],
+        [yearNow, monthNow, dateNow, dayNow, calendarType],
     );
 
     return (
